server-python/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.31.0
python-dotenv>=1.0.0
-e ./agora-agent-python-sdk
